[TRAINER-1768] Fixed broken tests (documentation still needs updating).
