+++ lib/kafka_message_bus/examples/sample_message_data_factory_implementation.ex
@@ -21,13 +21,13 @@ defmodule KafkaMessageBus.Examples.SampleMessageDataFactoryImplementation do
   on_create function. This is the default behavior for unrecognized message data
   types that are encountered on the bus.
   """
-  alias KafkaMessageBus.Examples.SampleMessageData2
+  alias KafkaMessageBus.Examples.SampleMessageData
   alias KafkaMessageBus.Messages.MessageData.UnrecognizedMessageDataType
   require Logger
 
   def on_create(%{} = data, "sample_resource", "sample_action") do
     Logger.info(fn -> "Creating for sample_resource and sample_action: #{inspect(data)}" end)
-    SampleMessageData2.new(data)
+    SampleMessageData.new(data)
   end
 
   use UnrecognizedMessageDataType, :on_create
